+++ js-library/src/services/ens.ts
@@ -1,7 +1,7 @@
 import { utils, Contract, Signer, providers, BigNumber } from 'ethers'
 import { ENS_DOMAIN, NULL_ADDRESS } from '../constants/constants'
 import { ENS_ENVIRONMENT_CONFIGS } from '../constants/environment'
-import { waitTransaction } from '../utils/tx'
+import { waitRequestTransaction } from '../utils/tx'
 import { joinPublicKey, isPublicKeyValid, splitPublicKey } from '../utils/keys'
 import { Environments } from '../model/environments.enum'
 import { EnsUserData } from '../model/ens-user-data.model'
@@ -13,6 +13,7 @@ import FDSRegistrarContractLocal from '../contracts/FDSRegistrar/FDSRegistrar.js
 import { Username } from '../model/domain.type'
 import { assertUsername } from '../utils/domains'
 import { checkMinBalance, extractMessageFromFailedTx, isTxError } from '../utils/blockchain'
+import { ServiceRequest } from '../model/service-request.model'
 
 const { keccak256, toUtf8Bytes, namehash } = utils
 
@@ -23,6 +24,19 @@ export const PublicResolverContract = PublicResolverContractLocal
 export const FDSRegistrarContract = FDSRegistrarContractLocal
 
 const MIN_BALANCE = BigNumber.from('10000000000000000')
+
+enum RegisterUsernameStage {
+  FDS_REGISTER_COMPLETED = 1,
+  SET_RESOLVER_COMPLETED = 2,
+  SET_PUBLIC_KEY_COMPLETED = 3,
+}
+
+export interface RegisterUsernameRequestData {
+  username: Username
+  address: EthAddress
+  publicKey: PublicKey
+  expires: number
+}
 
 /**
  * ENS Class
@@ -94,52 +108,141 @@ export class ENS {
   }
 
   /**
-   * Sets owner of the provided username on ENS
+   * Creates a request object that needs to be used to invoke the registerUsername method.
+   * The request object can be used to invoke the method multiple times without making any
+   * inconsistencies. This can be useful if registration fails because of insufficient funds,
+   * or some other reason.
    * @param username ENS username
    * @param address Owner address of the username
    * @param publicKey Hex string of a public key
+   * @returns ServiceRequest instance that can be used to invoke the request multiple times if fails.
    */
-  public async registerUsername(
+  public createRegisterUsernameRequest(
     username: Username,
     address: EthAddress,
     publicKey: PublicKey,
     expires: number = 86400,
-  ): Promise<void> {
+  ): ServiceRequest<RegisterUsernameRequestData> {
+    return {
+      stage: 0,
+      data: {
+        username,
+        address,
+        publicKey,
+        expires,
+      },
+      completedTxs: [],
+    }
+  }
+
+  /**
+   * Sets owner of the provided username on ENS
+   * @param registerRequest request object previously created by the createRegisterUsernameRequest method
+   */
+  public async registerUsername(registerRequest: ServiceRequest<RegisterUsernameRequestData>): Promise<void> {
+    const {
+      data: { username, address, publicKey, expires },
+    } = registerRequest
     try {
-      assertUsername(username)
+      if (registerRequest.stage < RegisterUsernameStage.FDS_REGISTER_COMPLETED) {
+        assertUsername(username)
 
-      let ownerAddress: EthAddress = NULL_ADDRESS
+        let ownerAddress: EthAddress = NULL_ADDRESS
 
-      if (this.config.performChecks) {
-        await checkMinBalance(this.provider, address, MIN_BALANCE)
+        if (this.config.performChecks) {
+          await checkMinBalance(this.provider, address, MIN_BALANCE)
 
-        ownerAddress = await this.getUsernameOwner(username)
+          ownerAddress = await this.getUsernameOwner(username)
 
-        if (ownerAddress !== NULL_ADDRESS) {
-          throw new Error(`ENS: Username ${username} is not available`)
+          if (ownerAddress !== NULL_ADDRESS) {
+            throw new Error(`ENS: Username ${username} is not available`)
+          }
         }
+
+        if (ownerAddress === NULL_ADDRESS) {
+          await waitRequestTransaction(this._provider, registerRequest, () =>
+            this._fdsRegistrarContract.register(keccak256(toUtf8Bytes(username)), address, expires),
+          )
+        }
+
+        registerRequest.stage = RegisterUsernameStage.FDS_REGISTER_COMPLETED
       }
 
-      if (ownerAddress === NULL_ADDRESS) {
-        await waitTransaction(
-          this._fdsRegistrarContract.register(keccak256(toUtf8Bytes(username)), address, expires),
+      if (registerRequest.stage < RegisterUsernameStage.SET_RESOLVER_COMPLETED) {
+        await waitRequestTransaction(this._provider, registerRequest, () =>
+          this._ensRegistryContract.setResolver(
+            this.hashUsername(username),
+            this._publicResolverContract.address,
+          ),
         )
+
+        registerRequest.stage = RegisterUsernameStage.SET_RESOLVER_COMPLETED
       }
 
-      await waitTransaction(
-        this._ensRegistryContract.setResolver(
-          this.hashUsername(username),
-          this._publicResolverContract.address,
-        ),
-      )
+      if (registerRequest.stage < RegisterUsernameStage.SET_PUBLIC_KEY_COMPLETED) {
+        await this.setPublicKey(registerRequest, username, publicKey)
 
-      await this.setPublicKey(username, publicKey)
+        registerRequest.stage = RegisterUsernameStage.SET_PUBLIC_KEY_COMPLETED
+      }
     } catch (error) {
       if (isTxError(error)) {
         throw new Error(extractMessageFromFailedTx(error))
       }
       throw error
     }
+  }
+
+  /**
+   * Estimates gas amount for the registerUsername method
+   * @param username ENS username
+   * @param address Owner address of the username
+   * @param publicKey Hex string of a public key
+   * @param customRpc (optional) custom RPC provider if the default one can't calculate gas
+   * @returns gas amount estimation
+   */
+  public async registerUsernameEstimateGas(
+    username: Username,
+    address: EthAddress,
+    publicKey: PublicKey,
+    expires: number = 86400,
+    customRpc?: providers.JsonRpcProvider,
+  ): Promise<BigNumber> {
+    const rpc = customRpc || this._provider
+    const [publicKeyX, publicKeyY] = splitPublicKey(publicKey)
+
+    const encodedRegisterFn = this._fdsRegistrarContract.interface.encodeFunctionData('register', [
+      keccak256(toUtf8Bytes(username)),
+      address,
+      expires,
+    ])
+
+    const encodedSetResolverFn = this._ensRegistryContract.interface.encodeFunctionData('setResolver', [
+      this.hashUsername(username),
+      this._publicResolverContract.address,
+    ])
+
+    const encodedSetPubkeyFn = this._publicResolverContract.interface.encodeFunctionData('setPubkey', [
+      this.hashUsername(username),
+      publicKeyX,
+      publicKeyY,
+    ])
+
+    const gasAmounts = await Promise.all([
+      rpc.estimateGas({
+        to: this._fdsRegistrarContract.address,
+        data: encodedRegisterFn,
+      }),
+      rpc.estimateGas({
+        to: this._ensRegistryContract.address,
+        data: encodedSetResolverFn,
+      }),
+      rpc.estimateGas({
+        to: this._publicResolverContract.address,
+        data: encodedSetPubkeyFn,
+      }),
+    ])
+
+    return gasAmounts.reduce((sum, amount) => sum.add(amount), BigNumber.from(0))
   }
 
   /**
@@ -173,13 +276,18 @@ export class ENS {
 
   /**
    * Sets user's public key to the user's ENS entry
+   * @param request object previously created by the createRegisterUsernameRequest method
    * @param username ENS username
    * @param publicKey Public key that will be added to ENS
    */
-  public setPublicKey(username: Username, publicKey: PublicKey): Promise<void> {
+  public setPublicKey(
+    request: ServiceRequest<RegisterUsernameRequestData>,
+    username: Username,
+    publicKey: PublicKey,
+  ): Promise<void> {
     assertUsername(username)
     const [publicKeyX, publicKeyY] = splitPublicKey(publicKey)
-    return waitTransaction(
+    return waitRequestTransaction(this._provider, request, () =>
       this._publicResolverContract.setPubkey(this.hashUsername(username), publicKeyX, publicKeyY),
     )
   }